# panel/openapi.yaml
openapi: "3.0.3"
info:
  title: SBoard Panel API
  version: "1.0.0"
  description: SBoard panel management API
servers:
  - url: /api
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    AdminProfile:
      type: object
      required:
        - username
      properties:
        username:
          type: string
...
    BootstrapStatus:
      type: object
      required:
        - needs_setup
      properties:
        needs_setup:
          type: boolean
    BootstrapResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
...
    BootstrapRequest:
      type: object
      required:
        - username
        - password
        - confirm_password
      properties:
        username:
          type: string
        password:
          type: string
        confirm_password:
          type: string
        setup_token:
          type: string
    UpdateAdminProfileRequest:
      type: object
      required:
        - old_password
      properties:
        new_username:
          type: string
        old_password:
          type: string
        new_password:
          type: string
        confirm_password:
          type: string
    UpdateSystemSettingsRequest:
      type: object
      properties:
        subscription_base_url:
          type: string
        timezone:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /admin/bootstrap:
    get:
      operationId: getBootstrapStatus
      summary: Check if admin setup is needed
      security: []
      responses:
        "200":
          description: Bootstrap status
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/BootstrapStatus"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: bootstrap
      summary: Create initial admin account
      security: []
      parameters:
        - name: X-Setup-Token
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapRequest"
      responses:
        "200":
          description: Admin created
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/BootstrapResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Auth ────────────────────────────────────────────────
...
  /admin/profile:
    get:
      operationId: getAdminProfile
      summary: Get admin profile
      responses:
        "200":
          description: Admin profile
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/AdminProfile"
        "428":
          description: Needs setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateAdminProfile
      summary: Update admin profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminProfileRequest"
      responses:
        "200":
          description: Updated admin profile
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/AdminProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "428":
          description: Needs setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Subscription ────────────────────────────────────────
  /sub/{user_uuid}:
    get:
      operationId: getSubscription
      summary: Get user subscription
      security: []
      parameters:
        - name: user_uuid
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum:
              - singbox
              - v2ray
      responses:
        "200":
          description: Subscription content
          content:
            application/json:
              schema:
                type: object
                description: sing-box configuration
            text/plain:
              schema:
                type: string
                description: v2ray subscription (base64 encoded links)
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Users ───────────────────────────────────────────────
...
  /traffic/nodes/summary:
    get:
      operationId: getTrafficNodesSummary
      summary: Traffic summary per node
      parameters:
        - name: window
          in: query
          description: "Time window (e.g. 24h, 7d, 30d, all)"
          schema:
            type: string
      responses:
        "200":
          description: Per-node traffic summary
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrafficNodeSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /traffic/total/summary:
    get:
      operationId: getTrafficTotalSummary
      summary: Total traffic summary
      parameters:
        - name: window
          in: query
          description: "Time window (e.g. 24h, 7d, 30d, all)"
          schema:
            type: string
      responses:
        "200":
          description: Total traffic summary
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/TrafficTotalSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /traffic/timeseries:
    get:
      operationId: getTrafficTimeseries
      summary: Traffic timeseries
      parameters:
        - name: window
          in: query
          description: "Time window (e.g. 24h, 7d, 30d, all)"
          schema:
            type: string
        - name: bucket
          in: query
          description: Time bucket granularity
          schema:
            type: string
            enum:
              - minute
              - hour
              - day
        - name: node_id
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Traffic timeseries
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrafficTimeseriesPoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── System ──────────────────────────────────────────────
...
  /system/settings:
    get:
      operationId: getSystemSettings
      summary: Get system settings
      responses:
        "200":
          description: System settings
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/SystemSettings"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateSystemSettings
      summary: Update system settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSystemSettingsRequest"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/SystemSettings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Sync Jobs ───────────────────────────────────────────
  /sync-jobs:
    get:
      operationId: listSyncJobs
      summary: List sync jobs
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - name: node_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
            enum:
              - queued
              - running
              - success
              - failed
        - name: trigger_source
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Sync job list
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SyncJobListItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /sync-jobs/{id}:
    get:
      operationId: getSyncJob
      summary: Get sync job detail
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Sync job detail with attempts
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/SyncJobDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...